+++ docs/releases.rst
@@ -6,6 +6,7 @@ Version 2.1, In Development
 
 * Use the SmartyPants library to generate better quotation
   marks for Markdown.
+* Composers can be forced to compose with the ``--force`` flag.
 
 Version 2.0, Released July 25, 2015
 -----------------------------------